Implement Level Object
- Implement the Level object that will store level information. This
information will be pulled from the XML data.
- Do not allow Account/Level sprites to be set directly with an
accessor to prevent various problems such as assigning an avatar that
doesn't actually point to the correct image location field that is
stored in the database. Overall it should provide better security and
reduce errors such as forgetting to change an image field when new
avatar is added to an account in the future.
- Remove Singletons avatars at the end of each scene. CCSprite can only
be assigned to one scene at a time. The previous solution to remove the
sprite from the scene wasn't adequate because it would make the sprite
disappear before the end of the scene transition.
- Add levels to the Singleton class used to indicate the currently
selected level that is being played. In the future it will be used for
restarting the level and selecting the next level.
- Fix most of the memory leaks in the LoginLayer.